# inflate: improve performance of memory copy operations