@Dead2 2017-03-17
This looks really good. I have not done any testing with it yet, but I read through the code-changes, and noticed nothing that seems objectionable.

Two things though;
- Can you add a copyright header to memcopy.h? I am assuming this is code that you wrote and that you can zlib license it.
- A small nitpick really, if you could change memcopy.h to use a line length of 120 chars, that would be great. Most people are not limited to 80 chars wide terminals any more, so we have changed zlib-ng to use 120 chars, since there are quite a few places where that makes the code easier to read.