@sebpop 2017-03-17
Hans, thank you for your fast review.
Would the following text be enough to be added to the header file?

/* memcopy.h -- inline functions to copy small data chunks.
 * For conditions of distribution and use, see copyright notice in zlib.h
 */

I am preparing an updated patch with your recommendations.


On Fri, Mar 17, 2017 at 7:29 AM, Hans Kristian Rosbach <
notifications@github.com> wrote:

> This looks really good. I have not done any testing with it yet, but I
> read through the code-changes, and noticed nothing that seems objectionable.
>
> Two things though;
>
>    - Can you add a copyright header to memcopy.h? I am assuming this is
>    code that you wrote and that you can zlib license it.
>    - A small nitpick really, if you could change memcopy.h to use a line
>    length of 120 chars, that would be great. Most people are not limited to 80
>    chars wide terminals any more, so we have changed zlib-ng to use 120 chars,
>    since there are quite a few places where that makes the code easier to read.
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/Dead2/zlib-ng/pull/93#issuecomment-287341618>, or mute
> the thread
> <https://github.com/notifications/unsubscribe-auth/AAisTXonMO2v5dCHQfzL6b5UZn1TbwMzks5rmnyMgaJpZM4MfhjF>
> .
>