@Dead2 2017-03-18
It was brought to my attention that this patchset probably does not remain compatible with big-endian machines. We should probably fix that, either by just using the original code for those, or adapting the new code to work on big-endian as well where possible.
@sebpop If you can, we'd love to see you in #zlib-ng on the freenode irc network.