@simhos01 2017-03-18
If you have Ubuntu, you might try testing big-endian [this way](https://gist.github.com/simhos01/5eda2d649cabba8cf5d50eca78cd386b).