@sebpop 2017-03-19
Thanks Hans for pointing out the issue with big-endian.
Thanks Simon for the instructions on how to test on qemu big-endian code.
I have access to several machines that are big-endian in the gcc farm:
https://gcc.gnu.org/wiki/CompileFarm
Let me investigate the problem and I will post an amended patch.