@sebpop 2017-03-19
The new patch passes on big-endian powerpc-linux.