@sebpop 2017-03-20
I think I addressed all the requested changes in the new patch.

I also did a comparison with Simon's "data privatization" by adding padding to the end of the window in order to remove some of the branches and checks that avoid runing over the allocated buffer.  This had a small effect on the performance, about 5% on PNG decompression.  Also those changes do not make possible the change to inflate.c (that works on a different buffer than the window in inffast.c) and the change to inflate.c accounts for about 20% speedup in a PNG decompression.  Overall, I think that adding the end of window padding complexifies the assumptions under which the memcpy operations are legal without much performance benefit.