@sebpop 2017-03-21
The last version of the patch passes on x86_64-linux and aarch64-linux with the extended set of tests that Mika suggested in https://github.com/sebpop/zlib-ng/commit/c37fd4195ae8846f1c35c31950edc9168d6b9173#commitcomment-21405774

I have also measured the performance of the change when used in a png and gzip decompression:

The number of executed instructions reported by valgrind --tool=cachegrind
on the decompression of a 50K PNG file on aarch64-linux:
- before the patch:
I   refs:      3,783,757,451
D   refs:      1,574,572,882  (869,116,630 rd   + 705,456,252 wr)

- with the patch:
I   refs:      2,569,373,984
D   refs:        908,492,140  (524,196,258 rd   + 384,295,882 wr)

The compression of a 260MB directory containing the code of llvm into a tar.gz
of 35MB and decompressing that with minigzip -d
on i7-4790K x86_64-linux, it takes 0.533s before the patch and 0.493s with the patch,
on Juno-r0 aarch64-linux A57, it takes 2.796s before the patch and 2.467s with the patch,
on Juno-r0 aarch64-linux A53, it takes 4.055s before the patch and 3.604s with the patch.