@RJVB 2017-03-21
On Tuesday March 21 2017 11:51:19 sebpop wrote:

> The compression of a 260MB directory containing the code of llvm into a tar.gz
> of 35MB 

Is that with or without the tar step? Creating the .tar is probably a not insignificant bottleneck.