@sebpop 2017-03-21
On Tue, Mar 21, 2017 at 2:10 PM, René Bertin <notifications@github.com>
wrote:
>
> Is that with or without the tar step? Creating the .tar is probably a not
> insignificant bottleneck.
>
>
The numbers are only for the inflate part: "minigzip -d test.tar.gz", no
tar, no compression.