@Dead2 2017-03-22
I had to make a difficult choice here, since your and simon's patches are so similar.
But it seems to me that simon's patch is a good middle-ground that your changes can mostly build upon or enhance. I hope you will be willing to re-base this on top of the current hacknslash6 branch.