@sebpop 2017-03-23
I have tested the performance and correctness of the new patch on x86_64-linux and ppc64-linux.  I will test aarch64-linux tomorrow when I will have access to the phones and Juno boards.