@mtl1979 2017-03-23
Ok for me when all the comments have been addressed... Still leaves some code that can be optimized further, like using do-while loops for less than 4 iterations. Like I mentioned earlier, those loops should be unrolled. It makes the code a little longer, but branch prediction in most current cpus can handle unrolled short loops better.