@simhos01 2017-03-23
I had imagined that the chunk copies I'd put in `inffast.c` would be hoisted out to `memcopy.h` by this patch, and that the preprocessor would choose between implementations for different architectures as appropriate.  My patch was conservative about touching any code that I wasn't going to test (ie., implementations for alignment-constrained or other cores I can't benchmark on) which left a lot of preprocessor mess inside `inffast.c`.  But I think that if we're bold (are we bold?) then it would be reasonable to throw away the preprocessor stuff and always make calls to equivalent inline functions.

The original zlib was clearly written with no faith in the compiler's ability to inline and optimise, but that was a long time ago and I think we can afford to lean on the compiler more.

The original bytewise copy loops are probably still the appropriate fallback implementation for architectures where we don't know if un-aligned load/store are available.  However, if they're abstracted into an inline functions then the non-overlapping version can take `restrict` pointers so the compiler might find the best optimisation for that architecture.

That's what I'd imagined.  What do others think?