@sebpop 2017-03-23
> using do-while loops for less than 4 iterations
Could you please be a bit more specific about what you mean by this and in which context you want to apply this transform?

If there are still small loops doing byte copies around it is because I was not able to get a better performance by rewriting those loops in another way.  Of course I have not tried all possible ways, so if you could point in the code which parts you would like me to rewrite in a different way, I can run some perf experiments to see whether that is worth changing the code.