@sebpop 2017-03-23
Simon, I like the idea of moving the code from inffast.c to memcopy.h: that would make the code in inffast.c more readable, and would move the complexity of selecting the best way to generate the memory operations to memcopy.h.  I will prepare a patch for this.